# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Vinyl",
    "description": "API consulta de discos",
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3200",
      "description": "Api Testes"
    },
    {
      "url": "https://vinyldisc.azurewebsites.net",
      "description": "Api Produção"
    }
  ],
  "paths": {
    "/auth": {
      "post": {
        "tags": ["Authentication"],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/disc": {
      "get": {
        "tags": ["Disc"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "idDisc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDiscs"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ListDiscs"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDiscs"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Disc"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListDiscs"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ListDiscs"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListDiscs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Disc"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "idDisc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Config": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "nullable": true
          },
          "quantityByPage": {
            "type": "integer",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "nullable": true
          },
          "hasNext": {
            "type": "boolean",
            "nullable": true
          },
          "hasPrevious": {
            "type": "boolean",
            "nullable": true
          },
          "quantityPages": {
            "type": "integer",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ListDiscs": {
        "type": "object",
        "properties": {
          "idDisc": {
            "type": "integer",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "boolean",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "currency",
            "nullable": true
          },
          "observation": {
            "type": "string",
            "nullable": true
          },
          "lyricsquantity": {
            "type": "integer",
            "nullable": true
          },
          "condition": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Auth": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}